feat: simplify sidenav header
